Move grid utilities to common folder
