+++ imod/common/utilities/regrid.py
@@ -15,13 +15,13 @@ from imod.common.interfaces.ipointdatapackage import IPointDataPackage
 from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.common.interfaces.isimulation import ISimulation
 from imod.common.statusinfo import NestedStatusInfo
+from imod.common.utilities.clip import clip_by_grid
 from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
 from imod.common.utilities.value_filters import is_valid
 from imod.mf6.auxiliary_variables import (
     expand_transient_auxiliary_variables,
     remove_expanded_auxiliary_variables_from_dataset,
 )
-from imod.common.utilities.clip import clip_by_grid
 from imod.schemata import ValidationError
 from imod.typing.grid import (
     GridDataArray,
