+++ imod/tests/test_mf6/test_utilities/test_clip_utilities.py
@@ -10,9 +10,9 @@ from shapely.testing import assert_geometries_equal
 
 import imod
 from imod.common.interfaces.ipackagebase import IPackageBase
-from imod.mf6 import HorizontalFlowBarrierResistance
 from imod.common.utilities.clip import clip_by_grid
-from imod.mf6.utilities.grid import broadcast_to_full_domain
+from imod.common.utilities.grid import broadcast_to_full_domain
+from imod.mf6 import HorizontalFlowBarrierResistance
 
 
 @pytest.fixture(scope="function")
